backrose/requirements.txt
asgiref==3.7.2
black==25.1.0
click==8.1.8
Django==5.0.1
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
flake8==7.1.2
freezegun==1.5.1
gunicorn==22.0.0
iniconfig==2.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.3.5
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2023.4
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.9.0
